make clean architecture
