+++ .dart_tool/package_config_subset
@@ -1,3 +1,7 @@
+animated_splash_screen
+2.12
+file:///C:/Users/anshr/AppData/Local/Pub/Cache/hosted/pub.dev/animated_splash_screen-1.3.0/
+file:///C:/Users/anshr/AppData/Local/Pub/Cache/hosted/pub.dev/animated_splash_screen-1.3.0/lib/
 args
 3.0
 file:///C:/Users/anshr/AppData/Local/Pub/Cache/hosted/pub.dev/args-2.5.0/
@@ -82,6 +86,10 @@ meta
 2.12
 file:///C:/Users/anshr/AppData/Local/Pub/Cache/hosted/pub.dev/meta-1.11.0/
 file:///C:/Users/anshr/AppData/Local/Pub/Cache/hosted/pub.dev/meta-1.11.0/lib/
+page_transition
+2.17
+file:///C:/Users/anshr/AppData/Local/Pub/Cache/hosted/pub.dev/page_transition-2.1.0/
+file:///C:/Users/anshr/AppData/Local/Pub/Cache/hosted/pub.dev/page_transition-2.1.0/lib/
 path
 3.0
 file:///C:/Users/anshr/AppData/Local/Pub/Cache/hosted/pub.dev/path-1.9.0/
