+++ pubspec.yaml
@@ -20,6 +20,8 @@ dependencies:
   flutter_svg: ^2.0.10+1
   get: ^4.6.6
   dropdown_button2: ^2.3.9
+  animated_splash_screen: ^1.3.0
+
 
 dev_dependencies:
   flutter_test:
